# api/api/openapi.yml
openapi: '3.0.3'
info:
  title: AusCERT 2023 - Going back to the basics of API security
  description: API documentation for a demo API.
  contact:
    name: Kyle Jackson
    email: kyle.jackson@forgeresponse.com
  version: 1.0.0
servers:
  - url: https://auscert.forgeresponse.com
components:
  schemas:
...
    Investigation:
      type: object
      properties:
        id:
          type: string
        assignedTo:
          type: string
        attachments:
          type: object
          properties:
            id:
              type: string
        createdBy:
          type: string
        createdAt:
          type: string
        comments:
          type: object
          properties:
            id:
              type: string
        description:
          type: string
        investigationId:
          type: string
        severity:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        title:
          type: string
        tlp:
          type: integer
    CreateInvestigationRequest:
      type: object
      properties:
        assignedTo:
          type: string
        description:
          type: string
        severity:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        templateId:
          type: string
        title:
          type: string
        tlp:
          type: integer
    UpdateInvestigationRequest:
      type: object
      properties:
        assignedTo:
          type: string
        description:
          type: string
        severity:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        title:
          type: string
        tlp:
          type: integer
    InvestigationTemplate:
      type: object
      properties:
        id:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
        description:
          type: string
        titlePrefix:
          type: string
        severity:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        tlp:
          type: integer
    CreateInvestigationTemplateRequest:
      type: object
      properties:
        description:
          type: string
        titlePrefix:
          type: string
        severity:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        tlp:
          type: integer
    UpdateInvestigationTemplateRequest:
      type: object
      properties:
        description:
          type: string
        titlePrefix:
          type: string
        severity:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        tlp:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        id:
          type: string
        message:
          type: string
        code:
          type: integer
        transactionId:
          type: string
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: Authorization
      in: header
security:
  - ApiKeyAuth: []
paths:
...
  /v1/investigations:
    get:
      description: Returns a list of investigations
      tags:
        - Investigations
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: Specify the number of items to return
      responses:
        '200':
          description: Successfully returned a list of investigations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Investigation'
        '404':
          description: No records found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      description: Creates a new investigation
      tags:
        - Investigations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvestigationRequest'
      responses:
        '200':
          description: Successfully created investigation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Investigation'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/investigations/{id}:
    parameters:
      - name: id
        in: path
        description: Id of the investigation
        required: true
        schema:
          type: string
    get:
      description: Returns the specified investigation
      tags:
        - Investigations
      responses:
        '200':
          description: Successfully returned investigation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Investigation'
        '404':
          description: No record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      description: Updates the specified investigation
      tags:
        - Investigations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInvestigationRequest'
      responses:
        '200':
          description: Successfully updated access key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Investigation'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      description: Deletes the specified investigation
      tags:
        - Investigations
      responses:
        '204':
          description: Successfully deleted investigation
        '404':
          description: No record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/investigations/template:
    get:
      description: Returns a list of investigation templates
      tags:
        - Investigation Templates
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: Specify the number of items to return
      responses:
        '200':
          description: Successfully returned a list of investigation templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvestigationTemplate'
        '404':
          description: No records found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      description: Creates a new investigation template
      tags:
        - Investigation Templates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvestigationTemplateRequest'
      responses:
        '200':
          description: Successfully created investigation template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestigationTemplate'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/investigations/template/{id}:
    parameters:
      - name: id
        in: path
        description: Id of the investigation template
        required: true
        schema:
          type: string
    get:
      description: Returns the specified investigation template
      tags:
        - Investigation Templates
      responses:
        '200':
          description: Successfully returned investigation template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestigationTemplate'
        '404':
          description: No record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      description: Updates the specified investigation template
      tags:
        - Investigation Templates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInvestigationTemplateRequest'
      responses:
        '200':
          description: Successfully updated access key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestigationTemplate'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      description: Deletes the specified investigation template
      tags:
        - Investigation Templates
      responses:
        '204':
          description: Successfully deleted investigation template
        '404':
          description: No record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...